Task sharing
The user can now obtain a link to a task by using the share functionality.
This link can be used by an unlimited number of people for up to 10 hours,
after which the link cannot be used anymore. 
When opening the link, the app opens, loading the data. If the task's subject does
not exist in the users subjects, it will ask the user whether they want to create it or
select a different one. Either way, they eventually end up on the CreateTaskPage with
all fields filled in with the shared task's data.

This works both with and without an account. Note that without an account, it will upload
said data to a remote server and keep it there for at least 10 hours.

+++ lib/pages/subjects/subjects_widget.dart
@@ -122,23 +122,26 @@ class _SubjectState extends State<_Subject> {
         ),
       ],
       functions: [
-            () => showDialog(
-          context: context,
-          builder: (context) => SubjectNotesDialog(
-            notes: widget.subject.notes,
-            subjectId: widget.subject.id,
-          ),
-        ),
-            () => Get.to(() => CreateSubjectPage(subjectToEdit: widget.subject)),
-            () => showConfirmationDialog(
-          context: context,
-          title: 'delete'.tr,
-          content: 'confirm_delete_subject'
-              .trParams({'name': widget.subject.name}),
-          cancelText: 'cancel_caps'.tr,
-          confirmText: 'delete_caps'.tr,
-          onConfirm: () => Database.I.deleteSubject(widget.subject.id),
-        )
+        () => showDialog(
+              context: context,
+              builder: (context) => SubjectNotesDialog(
+                notes: widget.subject.notes,
+                subjectId: widget.subject.id,
+              ),
+            ),
+        () => Get.to(() => CreateSubjectPage(
+              initialData: widget.subject,
+              editMode: true,
+            )),
+        () => showConfirmationDialog(
+              context: context,
+              title: 'delete'.tr,
+              content: 'confirm_delete_subject'
+                  .trParams({'name': widget.subject.name}),
+              cancelText: 'cancel_caps'.tr,
+              confirmText: 'delete_caps'.tr,
+              onConfirm: () => Database.I.deleteSubject(widget.subject.id),
+            )
       ],
       child: AnimatedContainer(
         duration: const Duration(milliseconds: 150),
